backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
playwright==1.42.0
sqlalchemy[asyncio]==2.0.29
pydantic==2.6.4
python-multipart==0.0.9
google-generativeai==0.4.1
python-dotenv==1.0.1
openpyxl==3.1.2
pandas==2.2.1
